:white_check_mark: [Test] 채팅 메시지 등록 internal API 테스트
